# bundle the JuliaHO implementation into a package