--- a/julia_highly_optimized/RelatedHO/Project.toml
+++ b/julia_highly_optimized/RelatedHO/Project.toml
@@ -1,3 +1,8 @@
+name = "RelatedHO"
+uuid = "ca9105aa-6a53-48fb-a280-7830b67cc50b"
+authors = []
+version = "0.1.0"
+
 [deps]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"