--- a/julia_highly_optimized/RelatedHO/src/RelatedHO.jl
+++ b/julia_highly_optimized/RelatedHO/src/RelatedHO.jl
@@ -1,5 +1,7 @@
 module RelatedHO
 
+export main
+
 using JSON3
 using StructTypes
 using Dates