This lets us take advantage of native precompilation that just happens one time when the package is installed, and then the compiled code is reused each time the benchmark is run, rather than recompiling for each benchmark run. 

The runtime results are the same, but this makes the process of generating the benchmarks itself much faster:

Before:
```
[| Julia HO | 17.89 ms  |]
Julia HO:

        Benchmark 1: julia --project related.jl
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 5 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 5 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 5 milliseconds
        Processing time (w/o IO): 4 milliseconds
          Time (mean ± σ):      2.462 s ±  0.042 s    [User: 2.445 s, System: 0.372 s]
          Range (min … max):    2.405 s …  2.517 s    10 runs
         
Julia HO:

        Benchmark 1: julia --project related.jl
        Processing time (w/o IO): 28 milliseconds
        Processing time (w/o IO): 29 milliseconds
        Processing time (w/o IO): 28 milliseconds
          Time (mean ± σ):      2.524 s ±  0.018 s    [User: 2.484 s, System: 0.385 s]
          Range (min … max):    2.512 s …  2.537 s    2 runs
         
Julia HO:

        Benchmark 1: julia --project related.jl
        Processing time (w/o IO): 67 milliseconds
        Processing time (w/o IO): 63 milliseconds
        Processing time (w/o IO): 70 milliseconds
          Time (mean ± σ):      2.803 s ±  0.038 s    [User: 2.683 s, System: 0.432 s]
          Range (min … max):    2.776 s …  2.830 s    2 runs
```

After:
```
[| Julia HO | 17.84 ms  |]
Julia HO:

        Benchmark 1: julia --project=RelatedHO -e "using RelatedHO; main()"
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 5 milliseconds
        Processing time (w/o IO): 5 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 5 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 4 milliseconds
        Processing time (w/o IO): 5 milliseconds
          Time (mean ± σ):     543.0 ms ±   8.6 ms    [User: 538.7 ms, System: 362.3 ms]
          Range (min … max):   532.7 ms … 559.8 ms    10 runs
         
Julia HO:

        Benchmark 1: julia --project=RelatedHO -e "using RelatedHO; main()"
        Processing time (w/o IO): 26 milliseconds
        Processing time (w/o IO): 26 milliseconds
        Processing time (w/o IO): 26 milliseconds
          Time (mean ± σ):     612.1 ms ±   2.3 ms    [User: 621.6 ms, System: 349.5 ms]
          Range (min … max):   610.4 ms … 613.7 ms    2 runs
         
Julia HO:

        Benchmark 1: julia --project=RelatedHO -e "using RelatedHO; main()"
        Processing time (w/o IO): 69 milliseconds
        Processing time (w/o IO): 67 milliseconds
        Processing time (w/o IO): 69 milliseconds
          Time (mean ± σ):     829.9 ms ±   1.9 ms    [User: 774.4 ms, System: 413.0 ms]
          Range (min … max):   828.5 ms … 831.2 ms    2 runs
```